Clean up constants, utils, APIs between AI Chat Lab and AI Chat
