+++ apps/src/aichat/constants.ts
@@ -1,11 +1,9 @@
-import {SaveType} from '@cdo/apps/aichatLab/types';
-import {EVENTS} from '@cdo/apps/metrics/AnalyticsConstants';
 import {AiChatModelIds} from '@cdo/generated-scripts/sharedConstants';
 import modelsJson from '@cdo/static/aichat/modelDescriptions.json';
 
 import type {ValueOf} from '../types/utils';
 
-import type {ModelDescription} from './types';
+import type {ModelDescription, ModelParameters} from './types';
 
 export const modelDescriptions: ModelDescription[] =
   modelsJson.filter(isValidDescription);
@@ -18,18 +16,7 @@ function isValidDescription(
   );
 }
 
-export const saveTypeToAnalyticsEvent: {[key in SaveType]: string} = {
-  updateChatbot: EVENTS.UPDATE_CHATBOT,
-  publishModelCard: EVENTS.PUBLISH_MODEL_CARD_INFO,
-  saveModelCard: EVENTS.SAVE_MODEL_CARD_INFO,
-};
-
 export const MAX_NAME_LENGTH = 15;
-
-export enum ModalTypes {
-  WARNING = 'warning',
-  TEACHER_ONBOARDING = 'teacherOnboarding',
-}
 
 export const RESET_CONVERSATION_CUSTOMIZATION_UPDATES = [
   'selectedModelId',
@@ -42,9 +29,6 @@ export const RESET_CONVERSATION_CUSTOMIZATION_UPDATES = [
 export const MAX_NUM_FILES = 5;
 export const MAX_FILE_SIZE_MB = 5;
 
-export const FAQ_LINK =
-  'https://support.code.org/hc/en-us/articles/30162711193741-AI-Chat-Lab-FAQ';
-
 export const AI_SETTINGS_SUPPORT_LINK =
   'https://support.code.org/hc/en-us/articles/42550900593677-AI-Settings';
 
@@ -55,3 +39,12 @@ export const AI_CHAT_NOT_AUTHORIZED_TEACHER =
   'You must be a verified teacher or sign in via Google, Microsoft, Facebook, or an LMS to use and assign this tool.';
 export const AI_CHAT_NOT_AUTHORIZED_STUDENT =
   'Your teacher has not enabled this tool. Check with your teacher if you think this is an error.';
+
+export const MODEL_PARAMETER_LABELS: {
+  [key in keyof ModelParameters]: string;
+} = {
+  selectedModelId: 'Selected model',
+  temperature: 'Temperature',
+  systemPrompt: 'System prompt',
+  retrievalContexts: 'Retrieval',
+};
