+++ apps/src/aichat/redux/thunks/initializeAiCustomizations.ts
@@ -3,11 +3,11 @@ import {
   LevelAichatSettings,
   Visibility,
 } from '@cdo/apps/aichatLab/types';
+import {validateModelId} from '@cdo/apps/aichatLab/utils';
 import {
   DEFAULT_VISIBILITIES,
   EMPTY_AI_CUSTOMIZATIONS,
 } from '@cdo/apps/aichatLab/views/modelCustomization/constants';
-import {validateModelId} from '@cdo/apps/aichatLab/views/modelCustomization/utils';
 import {EVENTS} from '@cdo/apps/metrics/AnalyticsConstants';
 import {getTypedKeys} from '@cdo/apps/types/utils';
 import {AppDispatch} from '@cdo/apps/util/reduxHooks';
