+++ apps/src/aichat/types/feedback.ts
@@ -1,0 +1,5 @@
+import {ValueOf} from '@cdo/apps/types/utils';
+import {AiChatTeacherFeedback} from '@cdo/generated-scripts/sharedConstants';
+
+// Convenience type for the values of AiChatTeacherFeedback
+export type FeedbackValue = ValueOf<typeof AiChatTeacherFeedback>;
