+++ apps/src/aichatLab/types/index.ts
@@ -1,2 +1,4 @@
 export * from './customizations';
 export * from './levelProperties';
+export * from './toxicity';
+export * from './viewState';
