+++ apps/src/aichatLab/types/toxicity.ts
@@ -1,6 +1,4 @@
-import {FIELDS_CHECKED_FOR_TOXICITY} from '@cdo/apps/aichatLab/views/modelCustomization/constants';
-import {ValueOf} from '@cdo/apps/types/utils';
-import {AiChatTeacherFeedback} from '@cdo/generated-scripts/sharedConstants';
+import {FIELDS_CHECKED_FOR_TOXICITY} from '../views/modelCustomization/constants';
 
 /** Response structure for the detect toxicity API */
 export interface DetectToxicityResponse {
@@ -17,6 +15,3 @@ export interface FlaggedField {
     details: {evaluation: string};
   };
 }
-
-// Convenience type for the values of AiChatTeacherFeedback
-export type FeedbackValue = ValueOf<typeof AiChatTeacherFeedback>;
