+++ apps/src/aichatLab/types/viewState.ts
@@ -1,0 +1,9 @@
+export enum ViewMode {
+  EDIT = 'edit-mode',
+  PRESENTATION = 'presentation-mode',
+}
+
+export enum ModalTypes {
+  WARNING = 'warning',
+  TEACHER_ONBOARDING = 'teacherOnboarding',
+}
