+++ apps/src/aichatLab/views/TeacherOnboardingModal.tsx
