+++ apps/src/aichatLab/views/onboarding-modal.module.scss
